migrate ComparisonProperties
