+++ splink/athena/athena_comparison_level_library.py
@@ -15,6 +15,36 @@ from .athena_base import (
 )
 
 
+class AthenaComparisonProperties(AthenaBase):
+    @property
+    def _exact_match_level(self):
+        return exact_match_level
+
+    @property
+    def _null_level(self):
+        return null_level
+
+    @property
+    def _else_level(self):
+        return else_level
+
+    @property
+    def _array_intersect_level(self):
+        return array_intersect_level
+
+    @property
+    def _distance_in_km_level(self):
+        return distance_in_km_level
+
+    @property
+    def _distance_function_level(self):
+        return distance_function_level
+
+    @property
+    def _levenshtein_level(self):
+        return levenshtein_level
+
+
 class null_level(AthenaBase, NullLevelBase):
     pass
