+++ splink/athena/athena_comparison_library.py
@@ -5,40 +5,7 @@ from ..comparison_library import (
     ExactMatchBase,
     LevenshteinAtThresholdsComparisonBase,
 )
-from .athena_base import (
-    AthenaBase,
-)
-from .athena_comparison_level_library import (
-    array_intersect_level,
-    distance_function_level,
-    distance_in_km_level,
-    else_level,
-    exact_match_level,
-    levenshtein_level,
-    null_level,
-)
-
-
-class AthenaComparisonProperties(AthenaBase):
-    @property
-    def _exact_match_level(self):
-        return exact_match_level
-
-    @property
-    def _null_level(self):
-        return null_level
-
-    @property
-    def _else_level(self):
-        return else_level
-
-    @property
-    def _array_intersect_level(self):
-        return array_intersect_level
-
-    @property
-    def _distance_in_km_level(self):
-        return distance_in_km_level
+from .athena_comparison_level_library import AthenaComparisonProperties
 
 
 class exact_match(AthenaComparisonProperties, ExactMatchBase):
@@ -50,7 +17,7 @@ class distance_function_at_thresholds(
 ):
     @property
     def _distance_level(self):
-        return distance_function_level
+        return self._distance_function_level
 
 
 class levenshtein_at_thresholds(
@@ -58,7 +25,7 @@ class levenshtein_at_thresholds(
 ):
     @property
     def _distance_level(self):
-        return levenshtein_level
+        return self._levenshtein_level
 
 
 class array_intersect_at_sizes(
