+++ splink/duckdb/duckdb_comparison_level_library.py
@@ -18,6 +18,49 @@ from .duckdb_base import (
 )
 
 
+# Class used to feed our comparison_library classes
+class DuckDBComparisonProperties(DuckDBBase):
+    @property
+    def _exact_match_level(self):
+        return exact_match_level
+
+    @property
+    def _null_level(self):
+        return null_level
+
+    @property
+    def _else_level(self):
+        return else_level
+
+    @property
+    def _datediff_level(self):
+        return datediff_level
+
+    @property
+    def _array_intersect_level(self):
+        return array_intersect_level
+
+    @property
+    def _distance_in_km_level(self):
+        return distance_in_km_level
+
+    @property
+    def _distance_function_level(self):
+        return distance_function_level
+
+    @property
+    def _levenshtein_level(self):
+        return levenshtein_level
+
+    @property
+    def _jaro_winkler_level(self):
+        return jaro_winkler_level
+
+    @property
+    def _jaccard_level(self):
+        return jaccard_level
+
+
 class null_level(DuckDBBase, NullLevelBase):
     pass
