+++ splink/spark/spark_comparison_library.py
@@ -8,59 +8,7 @@ from ..comparison_library import (
     JaroWinklerAtThresholdsComparisonBase,
     LevenshteinAtThresholdsComparisonBase,
 )
-from .spark_base import (
-    SparkBase,
-)
-from .spark_comparison_level_library import (
-    array_intersect_level,
-    datediff_level,
-    distance_function_level,
-    distance_in_km_level,
-    else_level,
-    exact_match_level,
-    jaccard_level,
-    jaro_winkler_level,
-    levenshtein_level,
-    null_level,
-)
-
-
-class SparkComparisonProperties(SparkBase):
-    @property
-    def _exact_match_level(self):
-        return exact_match_level
-
-    @property
-    def _null_level(self):
-        return null_level
-
-    @property
-    def _else_level(self):
-        return else_level
-
-    @property
-    def _array_intersect_level(self):
-        return array_intersect_level
-
-    @property
-    def _datediff_level(self):
-        return datediff_level
-
-    @property
-    def _distance_in_km_level(self):
-        return distance_in_km_level
-
-    @property
-    def _levenshtein_level(self):
-        return levenshtein_level
-
-    @property
-    def _jaro_winkler_level(self):
-        return jaro_winkler_level
-
-    @property
-    def _jaccard_level(self):
-        return jaccard_level
+from .spark_comparison_level_library import SparkComparisonProperties
 
 
 class exact_match(SparkComparisonProperties, ExactMatchBase):
@@ -72,7 +20,7 @@ class distance_function_at_thresholds(
 ):
     @property
     def _distance_level(self):
-        return distance_function_level
+        return self._distance_function_level
 
 
 class levenshtein_at_thresholds(
@@ -80,7 +28,7 @@ class levenshtein_at_thresholds(
 ):
     @property
     def _distance_level(self):
-        return levenshtein_level
+        return self._levenshtein_level
 
 
 class jaro_winkler_at_thresholds(
@@ -88,7 +36,7 @@ class jaro_winkler_at_thresholds(
 ):
     @property
     def _distance_level(self):
-        return jaro_winkler_level
+        return self._jaro_winkler_level
 
 
 class jaccard_at_thresholds(
@@ -96,7 +44,7 @@ class jaccard_at_thresholds(
 ):
     @property
     def _distance_level(self):
-        return jaccard_level
+        return self._jaccard_level
 
 
 class array_intersect_at_sizes(
